Added authentication for houses routes

+++ routes/housesRoutes.js
@@ -1,22 +1,33 @@
 import { Router } from 'express'
 import db from '../db.js'
+import jwt from 'jsonwebtoken'
+import { jwtSecret } from '../secrets.js'
 
 const router = Router()
 
 // CREATE
 router.post('/houses', async (req, res) => {
-  const {
-    location,
-    bedrooms,
-    bathrooms,
-    description,
-    price_per_night,
-    host_id
-  } = req.body
   try {
+    const token = req.cookies.jwt
+    if (!token) {
+      throw new Error('No token provided')
+    } else {
+      const decodedToken = jwt.verify(token, jwtSecret)
+      if (!decodedToken) {
+        throw new Error('Invalid authentication token')
+      }
+    }
+    const {
+      location,
+      bedrooms,
+      bathrooms,
+      description,
+      price_per_night,
+      host_id
+    } = req.body
     const insertion =
       await db.query(` INSERT INTO houses (location, bedrooms, bathrooms, description, price_per_night, host_id)
-    VALUES ('${location}', ${bedrooms}, ${bathrooms}, '${description}', ${price_per_night}, ${host_id})
+    VALUES ('${location}', ${bedrooms}, ${bathrooms}, '${description}', ${price_per_night}, ${decodedToken.user_id})
 RETURNING *`)
     res.json(insertion.rows[0])
   } catch (err) {
@@ -63,6 +74,15 @@ router.get('/houses/:houseId', async (req, res) => {
 router.patch('/houses/:houseId', async (req, res) => {
   let houseId = req.params.houseId
   try {
+    const token = req.cookies.jwt
+    if (!token) {
+      throw new Error('No token provided')
+    } else {
+      const decodedToken = jwt.verify(token, jwtSecret)
+      if (!decodedToken) {
+        throw new Error('Invalid authentication token')
+      }
+    }
     const readRows = await db.query(
       `SELECT * FROM houses WHERE house_id = ${houseId}`
     )
@@ -71,17 +91,19 @@ router.patch('/houses/:houseId', async (req, res) => {
       throw new Error(`No house found with ID ${houseId}`)
     }
     const readObj = readRows.rows[0]
+    if (decodedToken.user_id !== readObj.host_id) {
+      throw new Error('You are not authorized')
+    }
     let location = req.body.location || readObj.location
     let bedrooms = req.body.bedrooms || readObj.bedrooms
     let bathrooms = req.body.bathrooms || readObj.bathrooms
     let description = req.body.description || readObj.description
     let price_per_night = req.body.price_per_night || readObj.price_per_night
-    let host_id = req.body.host_id || readObj.host_id
     const updateRows = await db.query(
       `UPDATE houses
     SET location = '${location}', bedrooms = ${bedrooms},
     bathrooms = ${bathrooms}, description = '${description}',
-    price_per_night = ${price_per_night}, host_id = ${host_id}
+    price_per_night = ${price_per_night}, host_id = ${decodedToken.user_id}
     WHERE house_id = ${req.params.houseId}
     RETURNING *`
     )
@@ -99,12 +121,24 @@ router.patch('/houses/:houseId', async (req, res) => {
 router.delete('/houses/:houseId', async (req, res) => {
   let houseId = req.params.houseId
   try {
+    const token = req.cookies.jwt
+    if (!token) {
+      throw new Error('No token provided')
+    } else {
+      const decodedToken = jwt.verify(token, jwtSecret)
+      if (!decodedToken) {
+        throw new Error('Invalid authentication token')
+      }
+    }
     const readRows = await db.query(
       `SELECT * FROM houses WHERE house_id = ${houseId}`
     )
     const readResult = readRows.rows[0]
     if (readResult === undefined) {
       throw new Error(`No house found with ID ${houseId}`)
+    }
+    if (decodedToken.user_id !== readResult.host_id) {
+      throw new Error('You are not authorized')
     }
     const deleteRows = await db.query(`
     DELETE FROM houses WHERE house_id = ${req.params.houseId}
